+++ crates/px-configuration/src/lib.rs
@@ -1,0 +1,6 @@
+mod global;
+mod profile;
+mod project;
+
+// pub use global::*;
+pub use project::*;